+++ src/binreg_eigen.cpp
@@ -77,7 +77,7 @@ double calc_ll_binreg_eigenchol(const NumericMatrix& X, const NumericVector& y,
 
     } // end iterations
 
-    if(!converged) r_warning("binreg didn't converge");
+    if(!converged) r_warning("binary trait regression didn't converge: increase maxit or tol");
 
     return llik;
 }
@@ -149,7 +149,7 @@ double calc_ll_binreg_eigenqr(const NumericMatrix& X, const NumericVector& y,
 
     } // end iterations
 
-    if(!converged) r_warning("binreg didn't converge");
+    if(!converged) r_warning("binary trait regression didn't converge: increase maxit or tol");
 
     return llik;
 }
@@ -256,7 +256,7 @@ List fit_binreg_eigenqr(const NumericMatrix& X,
         curllik = llik;
     } // end iterations
 
-    if(!converged) r_warning("binreg didn't converge");
+    if(!converged) r_warning("binary trait regression didn't converge: increase maxit or tol");
 
     // now get coefficients, SEs, etc.
     List fit = fit_linreg_eigenqr(XX, z, true, qr_tol);