+++ NEWS.md
@@ -1,4 +1,4 @@
-## qtl2 0.21-5 (2019-07-18)
+## qtl2 0.21-6 (2019-08-05)
 
 ### Major changes
 
@@ -19,6 +19,9 @@
 
 - Added functions `n_founders()` and `founders()` for getting the
   number of founders and the founder strain names for a cross2 object.
+
+- `scan1()` now takes an `hsq` argument, so that the residual
+  heritability may be specified rather than calculated.
 
 ### Minor changes